Remove some warnings

+++ EverythingToolbar/Controls/SettingsControl.xaml.cs
@@ -106,7 +106,7 @@ namespace EverythingToolbar
             (menu.Items[selectedIndex] as MenuItem).IsChecked = false;
 
             int[] fastSortExceptions = { 9, 10, 17, 18 };
-            if (EverythingSearch.Instance.GetIsFastSort((uint)selectedIndex + 1) ||
+            if (EverythingSearch.Instance.GetIsFastSort(selectedIndex + 1) ||
                 fastSortExceptions.Contains(selectedIndex + 1))
             {
                 Properties.Settings.Default.sortBy = selectedIndex + 1;
